+++ ext/fs/lib.rs
@@ -9,7 +9,6 @@ pub use crate::interface::FileSystem;
 pub use crate::interface::FsDirEntry;
 pub use crate::interface::FsError;
 pub use crate::interface::FsFileType;
-pub use crate::interface::FsPermissions;
 pub use crate::interface::FsResult;
 pub use crate::interface::FsStat;
 pub use crate::interface::OpenOptions;
@@ -17,10 +16,47 @@ use crate::ops::*;
 
 pub use crate::std_fs::StdFs;
 
+use deno_core::error::AnyError;
 use deno_core::OpState;
+use deno_core::Resource;
 use std::cell::RefCell;
 use std::convert::From;
+use std::path::Path;
 use std::rc::Rc;
+
+pub trait FsPermissions {
+  fn check_read(&mut self, p: &Path, api_name: &str) -> Result<(), AnyError>;
+  fn check_read_all(&mut self, api_name: &str) -> Result<(), AnyError>;
+  fn check_read_blind(
+    &mut self,
+    p: &Path,
+    display: &str,
+    api_name: &str,
+  ) -> Result<(), AnyError>;
+  fn check_write(&mut self, p: &Path, api_name: &str) -> Result<(), AnyError>;
+  fn check_write_all(&mut self, api_name: &str) -> Result<(), AnyError>;
+  fn check_write_blind(
+    &mut self,
+    p: &Path,
+    display: &str,
+    api_name: &str,
+  ) -> Result<(), AnyError>;
+
+  fn check(
+    &mut self,
+    open_options: &OpenOptions,
+    path: &Path,
+    api_name: &str,
+  ) -> Result<(), AnyError> {
+    if open_options.read {
+      self.check_read(path, api_name)?;
+    }
+    if open_options.write || open_options.append {
+      self.check_write(path, api_name)?;
+    }
+    Ok(())
+  }
+}
 
 struct UnstableChecker {
   pub unstable: bool,
@@ -52,6 +88,7 @@ pub(crate) fn check_unstable2(state: &Rc<RefCell<OpState>>, api_name: &str) {
 deno_core::extension!(deno_fs,
   deps = [ deno_web ],
   parameters = [Fs: FileSystem, P: FsPermissions],
+  bounds = [Fs::File: Resource],
   ops = [
     op_cwd<Fs, P>,
     op_umask<Fs>,
